+++ lua/cokeline/augroups.lua
@@ -10,7 +10,7 @@ local toggle = function()
 end
 
 ---@param bufnr  number
-local focus = function(bufnr)
+local focus_on_delete = function(bufnr)
   if #_G.cokeline.valid_buffers == 1 then return end
 
   local deleted_buffer = vim_filter(
@@ -19,29 +19,21 @@ local focus = function(bufnr)
   )[1]
   if not deleted_buffer then return end
 
-  ---@param buffer  Buffer
-  ---@return Buffer
-  local prev = function(buffer)
-    return _G.cokeline.valid_buffers[
-      buffer._valid_index ~= 1
-      and buffer._valid_index - 1
-       or 2
-    ]
-  end
+  local prev_buffer = _G.cokeline.valid_buffers[
+    deleted_buffer._valid_index ~= 1
+    and deleted_buffer._valid_index - 1
+     or 2
+  ]
 
-  ---@param buffer  Buffer
-  ---@return Buffer
-  local next = function(buffer)
-    return _G.cokeline.valid_buffers[
-      buffer._valid_index ~= #_G.cokeline.valid_buffers
-      and buffer._valid_index + 1
-       or #_G.cokeline.valid_buffers - 1
-    ]
-  end
+  local next_buffer = _G.cokeline.valid_buffers[
+    deleted_buffer._valid_index ~= #_G.cokeline.valid_buffers
+    and deleted_buffer._valid_index + 1
+     or #_G.cokeline.valid_buffers - 1
+  ]
 
   vim.cmd(('b%s'):format(
-    (gl_settings.focus_on_delete == 'prev' and prev(deleted_buffer).number)
-    or (gl_settings.focus_on_delete == 'next' and next(deleted_buffer).number)
+    (gl_settings.focus_on_delete == 'prev' and prev_buffer.number)
+    or (gl_settings.focus_on_delete == 'next' and next_buffer.number)
   ))
 end
 
@@ -59,7 +51,7 @@ local setup = function(settings)
     augroup cokeline_focus
       autocmd BufUnload *
         \ exe printf(
-        \   'lua require("cokeline/augroups").focus(%s)', expand('<abuf>')
+        \   'lua require("cokeline/augroups").focus_on_delete(%s)', expand('<abuf>')
         \ )
     augroup END
     ]])
@@ -67,7 +59,7 @@ local setup = function(settings)
 end
 
 return {
+  focus_on_delete = focus_on_delete,
   toggle = toggle,
-  focus = focus,
   setup = setup,
 }